Privatize custom iterators and generalize to all/node/element
I have moved a bit the methods, now the first is always the array like generic method, the second one is for nodes and the third one for elements. This is applied to iterator, stream and iterable utility methods. Also I have changed the iterator implementation using the nextSibling/nextElementSibling property, this should be more remove friendly (I think), so not both iterable support removal.