Apply Elements.body and unify Application/TodoItem/Footer asElement implementation
Although it is unnecessary, creating the root element always in the constructor is a good practice, this actually looks like a poor web component, so maybe in the future all this IsElement can be turned into components with some minor refactoring. Also it is a not-required/but-good-idea practice to separate event bindings from element creation. This facilitates a more clean definition of events behaviour and makes disappear some problems like referencing a parent element (because java do not allow to reference a field inside the field definition itself) and event handlers (bc it is pretty ugly if the builder returns the handler instead of the builder itself).

Also change the gwt maven config bc the rc7 changes make gwt-lib a bit broken, and the current version need to be nowhere to actually maven use the local source, if maven is able to resolve any artifact it will use the artifact instead of local sources, this plus this project uses stable version's instead of snapshots make it super easy to end up using obsolete sources.