Add elemental2 inherits
This is mandatory or GWT will fail to compile. Usually hidden bc if the
user uses elemento it almost always has already inherited elemental2,
but I noticed this bc webstorage is not so often used, and as elemento
has included its events, it will require this module to be inherited.
So, just add the modules that elemento-core actually uses.