Fix #46: Internet Explorer JavaScriptException when using template expressions
The `createTreeWalker` specification documents an optional third argument,
`filter`, which is described as a `NodeFilter` object containing an
`acceptNode` method. The Elemental2 implementation of `createTreeWalker` follows
this specification.

However, Internet Explorer expects the `filter` argument to be a function,
not an object with an `acceptNode` method. When an object is passed in instead,
Internet Explorer throws a JavaScriptException on the first call to the
`nextNode()` method of the tree walker object.

The workaround in this commit passes null as the `filter` argument and handles
the filtering in the subsequent while loop after each call to `nextNode()`.